Cleaned up code to simplify main.rs
