+++ src-tauri/src/mesh/serial_connection.rs
@@ -9,6 +9,7 @@ use std::{
 use prost::Message;
 use rand::{distributions::Standard, prelude::Distribution, Rng};
 use serialport::SerialPort;
+use tauri::{self, Manager};
 use tokio::sync::broadcast;
 
 use app::protobufs;
@@ -34,13 +35,20 @@ pub trait MeshConnection {
     fn send_raw(&mut self, data: Vec<u8>) -> Result<(), Box<dyn Error>>;
     fn write_to_radio(port: &mut Box<dyn SerialPort>, data: Vec<u8>) -> Result<(), Box<dyn Error>>;
 
-    // fn handle_from_radio(&mut self, data: Vec<u8>);
-    // fn handle_mesh_packet(&mut self, mesh_packet: protobufs::MeshPacket);
-    // fn handle_data_packet(
-    //     &mut self,
-    //     mesh_packet: protobufs::MeshPacket,
-    //     data_packet: protobufs::Data,
-    // );
+    fn dispatch_packet(
+        handle: tauri::AppHandle,
+        variant: app::protobufs::from_radio::PayloadVariant,
+    ) -> Result<(), Box<dyn Error>>;
+
+    fn handle_mesh_packet(
+        handle: tauri::AppHandle,
+        packet: protobufs::MeshPacket,
+    ) -> Result<(), Box<dyn Error>>;
+
+    fn handle_decoded_mesh_packet(
+        handle: tauri::AppHandle,
+        data: protobufs::Data,
+    ) -> Result<(), Box<dyn Error>>;
 
     fn generate_rand_id<T>() -> T
     where
@@ -135,6 +143,141 @@ impl MeshConnection for SerialConnection {
         let binding = [&magic_buffer, packet_slice].concat();
         let message_buffer: &[u8] = binding.as_slice();
         port.write(message_buffer)?;
+
+        Ok(())
+    }
+
+    fn dispatch_packet(
+        handle: tauri::AppHandle,
+        variant: app::protobufs::from_radio::PayloadVariant,
+    ) -> Result<(), Box<dyn Error>> {
+        match variant {
+            protobufs::from_radio::PayloadVariant::Channel(c) => {
+                // println!("Channel data: {:#?}", c);
+                handle.emit_all("channel", c)?;
+            }
+            protobufs::from_radio::PayloadVariant::Config(c) => {
+                // println!("Config data: {:#?}", c);
+                handle.emit_all("config", c)?;
+            }
+            protobufs::from_radio::PayloadVariant::ConfigCompleteId(c) => {
+                // println!("Config complete id data: {:#?}", c);
+                handle.emit_all("config_complete", c)?;
+            }
+            protobufs::from_radio::PayloadVariant::LogRecord(l) => {
+                // println!("Log record data: {:#?}", l);
+                handle.emit_all("log_record", l)?;
+            }
+            protobufs::from_radio::PayloadVariant::ModuleConfig(m) => {
+                // println!("Module config data: {:#?}", m);
+                handle.emit_all("module_config", m)?;
+            }
+            protobufs::from_radio::PayloadVariant::MyInfo(m) => {
+                // println!("My node info data: {:#?}", m);
+                handle.emit_all("my_node_info", m)?;
+            }
+            protobufs::from_radio::PayloadVariant::NodeInfo(n) => {
+                // println!("Node info data: {:#?}", n);
+                handle.emit_all("node_info", n)?;
+            }
+            protobufs::from_radio::PayloadVariant::Packet(p) => {
+                // println!("Packet data: {:#?}", p);
+                SerialConnection::handle_mesh_packet(handle, p)?;
+            }
+            protobufs::from_radio::PayloadVariant::Rebooted(r) => {
+                // println!("Rebooted data: {:#?}", r);
+                handle.emit_all("reboot", r)?;
+            }
+        };
+
+        Ok(())
+    }
+
+    fn handle_mesh_packet(
+        handle: tauri::AppHandle,
+        packet: protobufs::MeshPacket,
+    ) -> Result<(), Box<dyn Error>> {
+        let variant = packet.payload_variant.ok_or("No payload variant")?;
+
+        match variant {
+            protobufs::mesh_packet::PayloadVariant::Decoded(d) => {
+                println!("Decoded: {:#?}", d);
+                SerialConnection::handle_decoded_mesh_packet(handle, d)?;
+            }
+            protobufs::mesh_packet::PayloadVariant::Encrypted(e) => {
+                eprintln!("Encrypted packets not yet supported in Rust: {:#?}", e);
+            }
+        }
+
+        Ok(())
+    }
+
+    fn handle_decoded_mesh_packet(
+        handle: tauri::AppHandle,
+        data: protobufs::Data,
+    ) -> Result<(), Box<dyn Error>> {
+        match data.portnum() {
+            protobufs::PortNum::AdminApp => {
+                eprintln!("Admin application not yet supported in Rust");
+            }
+            protobufs::PortNum::AtakForwarder => {
+                eprintln!("ATAK forwarder not yet supported in Rust");
+            }
+            protobufs::PortNum::AudioApp => {
+                eprintln!("Audio app not yet supported in Rust");
+            }
+            protobufs::PortNum::IpTunnelApp => {
+                eprintln!("IP tunnel app not yet supported in Rust");
+            }
+            protobufs::PortNum::NodeinfoApp => {
+                eprintln!("Node info app not yet supported in Rust");
+            }
+            protobufs::PortNum::PositionApp => {
+                handle.emit_all("position", data)?;
+            }
+            protobufs::PortNum::PrivateApp => {
+                eprintln!("Private app not yet supported in Rust");
+            }
+            protobufs::PortNum::RangeTestApp => {
+                eprintln!("Range test app not yet supported in Rust");
+            }
+            protobufs::PortNum::RemoteHardwareApp => {
+                eprintln!("Remote hardware app not yet supported in Rust");
+            }
+            protobufs::PortNum::ReplyApp => {
+                eprintln!("Reply app not yet supported in Rust");
+            }
+            protobufs::PortNum::RoutingApp => {
+                handle.emit_all("routing", data)?;
+            }
+            protobufs::PortNum::SerialApp => {
+                eprintln!("Serial app not yet supported in Rust");
+            }
+            protobufs::PortNum::SimulatorApp => {
+                eprintln!("Simulator app not yet supported in Rust");
+            }
+            protobufs::PortNum::StoreForwardApp => {
+                eprintln!("Store forward packets not yet supported in Rust");
+            }
+            protobufs::PortNum::TelemetryApp => {
+                handle.emit_all("telemetry", data)?;
+            }
+            protobufs::PortNum::TextMessageApp => {
+                handle.emit_all("text", data)?;
+            }
+            protobufs::PortNum::TextMessageCompressedApp => {
+                eprintln!("Compressed text data not yet supported in Rust");
+            }
+            protobufs::PortNum::WaypointApp => {
+                eprintln!("Waypoint app not yet supported in Rust");
+            }
+            protobufs::PortNum::ZpsApp => {
+                eprintln!("ZPS app not yet supported in Rust");
+            }
+            _ => {
+                eprintln!("Unknown packet received");
+            }
+        }
 
         Ok(())
     }
